refactor: move requests to internal
