+++ services/user/register.go
@@ -1,7 +1,7 @@
 package user
 
 import (
-	"echo-demo-project/requests"
+	"echo-demo-project/internal/requests"
 	"echo-demo-project/server/builders"
 
 	"golang.org/x/crypto/bcrypt"
