+++ tests/post_test.go
@@ -2,17 +2,18 @@ package tests
 
 import (
 	"database/sql/driver"
-	"echo-demo-project/requests"
+	"echo-demo-project/internal/requests"
 	"echo-demo-project/server"
 	"echo-demo-project/server/handlers"
 	"echo-demo-project/services/token"
 	"echo-demo-project/tests/helpers"
+	"net/http"
+	"testing"
+
 	"github.com/DATA-DOG/go-sqlmock"
 	"github.com/golang-jwt/jwt/v5"
 	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/assert"
-	"net/http"
-	"testing"
 )
 
 const postId = "1"
